+++ static_tests/run.sh
@@ -23,11 +23,13 @@ test_path() {
     echo "Check if expected errors are deteced ..."
     output=$(mypy --no-incremental static_tests/path.py)
 
-    run_check "Line 12"; echo "$output" | grep "static_tests/path.py:12:" >/dev/null || check_fail; check_ok
-    run_check "Line 18"; echo "$output" | grep "static_tests/path.py:18:" >/dev/null || check_fail; check_ok
-    run_check "Line 26"; echo "$output" | grep "static_tests/path.py:26:" >/dev/null || check_fail; check_ok
+    run_check "Line 13"; echo "$output" | grep "static_tests/path.py:13:" >/dev/null || check_fail; check_ok
+    run_check "Line 19"; echo "$output" | grep "static_tests/path.py:19:" >/dev/null || check_fail; check_ok
+    run_check "Line 27"; echo "$output" | grep "static_tests/path.py:27:" >/dev/null || check_fail; check_ok
+    run_check "Line 30"; echo "$output" | grep "static_tests/path.py:30:" >/dev/null || check_fail; check_ok
+    run_check "Line 33"; echo "$output" | grep "static_tests/path.py:33:" >/dev/null || check_fail; check_ok
 
-    run_check "Ensure there were no other errors"; [[ $(echo "$output" | wc -l) = "3" ]] || check_fail; check_ok
+    run_check "Ensure there were no other errors"; [[ $(echo "$output" | wc -l) = "5" ]] || check_fail; check_ok
 }
 
 test_testcase() {